taintfox: adding URL taint propagation
Adding taint propagation to the nsStandardURL so that taint
is propagated correctly when a normalized URL is created.

Propagates taint for the nsIURI xpcom interface as all
properties are stored as strings.

Some small additions to the URL class to copy taint so
that Javascript URL objects are also taint aware.