Foxhound: second attempt to fix JsonPath crashes
It seems creating JSONPath strings on the Tenured heap worked
until a sweep, and then similar segfaults occured.

This fix does two things: 1) create the JSON path strings as atoms,
which seems to be more stable, and 2) wrap them as rooted strings,
so they are collected in GC sweeps and don't leak memory.