Fixes Node.normalize() losing taints
If a DOM node has several text child nodes that are direct siblings, it is possible to merge them by calling [normalize()](https://developer.mozilla.org/en-US/docs/Web/API/Node/normalize). Prior to this change only the taint status of the first text node was considered and the remaining text nodes lost theirs. This PR fixes this and adds a test case.